requirements.txt
# Dependências principais do Flask
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.9.0

# Banco de dados (adicionado)
mysql-connector-python==8.2.0  # Versão mais recente com melhor compatibilidade
PyMySQL==1.1.0  # Conector alternativo para caso de problemas com o mysql-connector

# Segurança
bcrypt==4.0.1
python-dotenv==1.0.0

# Performance
Flask-Caching==2.0.2
cachelib==0.9.0

# Requisições HTTP
requests==2.31.0
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10

# Opcional para futuro uso
# SQLAlchemy==2.0.20
# Flask-SQLAlchemy==3.0.5
# Flask-Migrate==4.0.4